Send TERM rather than QUIT to Puma cluster
Sending `QUIT` to the Puma cluster brutally murders it, leading to
workers dying off with `exit!` when they detect the parent has died.  In
turn, this causes `at_exit` handlers to be skipped.

This patch changes `QUIT` for `TERM` for a graceful shutdown:

> `TERM`: send `TERM` to worker. Worker will attempt to finish then
> exit.

Aside: it could equally use `INT`, but `TERM` seems more descriptive
even if it may cause signal handlers to run multiple times as `TERM` is
already sent by the dyno manager.  However, code should deal with
multiple entries to signal handlers anyway, as the dyno manager does not
guarantee that `TERM` is not sent multiple times in any case!

This commit also provides for specification of the signal to use via an
option to the backend: `term_signal`

While we're at it, disable obsolete rubies from CI and add 2.1.x and 2.2.x
series.